Restructure ccd api module
