+++ src/main/java/uk/gov/hmcts/reform/bulkscanccdeventhandler/ccd/api/model/Event.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.bulkscanccdeventhandler.ccd;
+package uk.gov.hmcts.reform.bulkscanccdeventhandler.ccd.api.model;
 
 public class Event {
     public final String id;
